Add diagnosis to a banff dataset

This function takes in input a dataset which is given for all the observation a diagnosis. To be able to perform, the dataset must not contain any errors that banff_launcher() would not have detected. Please prefer using banff_launcher() to effectuate additional tests.

add_diagnosis(banff_dataset)

Arguments

banff_dataset

A tibble object.

Value

A tibble object which contain additional columns refering to diagnosiss.

Examples

{

library(fabR)
banff_file <- system.file("extdata", "example.xlsx", package = "lakebanff")
banff_dataset <- read_excel_allsheets(banff_file)
add_diagnosis(banff_dataset)

}
#> # A tibble: 0 × 158
#> # ℹ 158 variables: patient_id <lgl>, date_tx <lgl>, Sex <lgl>, Age_tx <lgl>,
#> #   Ethnicity <lgl>, Cause.of.ESKD <lgl>, First.Dialysis.Date <lgl>,
#> #   Dialysis.Modality <lgl>, Cold.Ischemia.Time <lgl>, sc_recip_dna <lgl>,
#> #   sc_don_dna <lgl>, Donor_origin <lgl>, Donor_ID <lgl>, don_type <lgl>,
#> #   Donor.Age <lgl>, Donor.Sex <lgl>, biopsy_id <lgl>, sc_date_bx <lgl>,
#> #   type_bx <lgl>, adequacy <lgl>, TCMR.patho <lgl>, ABMR.patho <lgl>,
#> #   Fixation <lgl>, microscopy <lgl>, Grade <lgl>, c4d_if <lgl>, …