Main function to start the program

This function takes in input a path string of an input file. The function internally adresses series of tests that assess the input dataset. If any of these tests fails, the user gets information allowing them to correct the file and reprocess. Once the tests do not fail, the dataset is given for all the observation a diagnosis (using internally the function add_diagnosis()). The output dataset, along with its associated labels (label:en by defaut) are provided to the user in an Excel format file accessibee in the output_folder they provided (working direct as a default.) The dataset comes with a report that summarize information about variable distributions and descriptive statistics.

banff_launcher(
  banff_file,
  output_folder = getwd(),
  language = "label:en",
  option_filter,
  detail = FALSE,
  include_banff_dictionary = FALSE
)

Arguments

banff_file

A character string identifying the name of the file to be processed.

output_folder

A character string identifying the folder path where the bookdown report files will be saved.

language

optional input allowing user to get their diagnosis in a specific language. Options are "label:en" (default), "label:fr", "label:de", "label:sp", "label:nl", "label:jp", "label:in".

option_filter

Optional argument allowing the user to filter its dataset using dplyr::filter() synthax. The column used to filter must exist.

detail

Whether the output should include temporary variables generated in the process or not. FALSE by default.

include_banff_dictionary

Whether the folder output should include the data dictionary (Excel format).

Value

Nothing to be returned. The function generates a folder (if not already exists) where Excel files are stored.

Examples

{

banff_file <- system.file("extdata", "example.xlsx", package = "lakebanff")
banff_launcher(banff_file, output_folder = tempdir())

}
#> Welcome to the banff diagnosis helper.
#> 
#> Press [enter] to continue or [esc] to quit.
#> 
#> 
#> [1/6] - Evaluation of input dataset 
#> [3/6] - add diagnosis to 'example'
#> [4/6] - generate labels for each column of 'example'
#> 
#> [5/6] - Assessment of the the output file 'example'
#> [6/6] - Export XLSX files in 'RtmpyGADhw'
#> C:\Users\guill\AppData\Local\Temp\RtmpyGADhw/example-2024-03-27-101436-diagnosiss.xlsx
#> C:\Users\guill\AppData\Local\Temp\RtmpyGADhw/example-2024-03-27-101436-report.xlsx
#> 
#> 
#> The assessment and addition of diagnosis on your dataset is now finished.
#> In the output folder, you will find an Excel file containing both labels and codes
#> of your dataset and for each participant, additional columns corresponding to your
#> diagnosis.
#> 
#> An assessment and summary report has been generated in your output folder to help you
#> assessing and interpret your dataset. You can open it as an Excel file.
#> 
#> For further information please refer to documentation